Make colorbar ticks format in generics plot configurable
Fix #1828
